Start implementation of DotnetInstaller
Copilot prompt:

Implement the GetConfiguredInstallType method in #file:'DotnetInstaler.cs' .

This method should look in the PATH and resolve "dotnet" or "dotnet.exe", depending on the current OS.  If it is not found, the return value should be SdkInstallType.None.  If it is found in Program Files, the install type should be SdkInstallType.Admin.  If it is found in another folder, the install type should be SdkInsntallType.User.

However, the method should also check the value of DOTNET_ROOT environment variable.  For a user install, DOTNET_ROOT should be set to the folder where the dotnet executable is found.  For an admin install, DOTNET_ROOT should not be set, but if it is set to the install path under program files, that is OK.  If the DOTNET_ROOT value doesn't match the install location, the method should return SdkInstallType.Inconsistent.