Add class for global.json contents
Copilot prompts:

Create a class to represent the contents of a global.json file, and create a corresponding JsonSerializerContext type.

Can you refer to https://learn.microsoft.com/en-us/dotnet/core/tools/global-json to update the schema you are using for the global.json file?  AllowPrerelease should be a booleann, and instead of Path there should be a Paths property which is an array of strings.