Fix check for dotnet folder in PATH
Copilot prompt:

In #method:'Microsoft.DotNet.Tools.Bootstrapper.DotnetInstaller.ConfigureInstallType':4481-6459 what I meant by removing a folder from the path if it has dotnet in it was that you should check the contents of each folder and if it is a dotnet installation folder then it should be removed from the path.  A simple way to check if it's a dotnet installation folder is if it has a dotnet executable in it.