Add Base Interface for Install
I renamed the DotnetInstaller in an aim to refactor the logic out some more. I think we could have the controller logic that talks to the CLI / UI layer be separate from the class that does the actual install logic to better follow the single responsibility principle and simplify the code into more pieces.