Only run the test templates on x64
This test step was added to ensure we scanned the packages pulled down in the test templates so it specifically had to run in the context of the CI machine rather than on helix. This causes issues when cross compiling on windows.

Since we only care about version, it should be fine to only run on x64.