[Blazor] Remove unnecessary update to the Blazor webassembly js file (#50949)
Hosted webassembly apps failed to publish due to a bug in how we were trying to update the blazor.webassembly.js asset definition. We realized that we did not have to do so, but that there was a bug precluding us from finding the asset during the publish process.

The fix corrects the identity of the blazor.webassembly.js to account for the fingerprint for the file we introduced in 10.0, which wasn't being taken into account.