Re-enable telemetry for global.json state - set error writer to swallow expected error messages (#51089)
When getting the global.json state for telemetry, we avoid the work to actually resolve an SDK. However, this means that the actual SDK resolution (`hostfxr_resolve_sdk2`) fails, resulting in messages printed to stderr by default. This change uses `hostfxr_set_error_writer` to swallow the expected errors for the duration of that call.