Have copilot handle formatting better
Updated guidelines for committing changes and testing.
dotnet format doesn't allow for formatting a single file. If copilot tries to format, it'll try the file, then the solution, and then the project resulting in lots of formatting. This will enable us to fix formatting but not mess up the diffs.