Add basic yml for running PR tests in dnup
we will want to consolidate this into the SDK later.

run pr ci on dnup

I don't have permission to add a pipeline so let's try to filter them out

I wish there was a built in method for yml to do this instead. For example, pr: has exclude which lets us not run the ci if changes are only in certain components. Ideally we could do something similar for the stages and condition them based on dnup/ non dnup changes? Note that we still need to have the .vsts-pr.yml file as the root pipeline file, and we can't just exclude running the entire thing as it controls the SDK pipeline as well.