Don't use global/ in the mutex
there is no concept of a global\ mutex prefix on unix which could be incorrect. On windows, this makes it so the mutex is not for the local user session but instead across all users. I'm not sure if this is needed but since we edit the system path, it might be. I will have to think about whether there will be issues with a higher-level admin user that has held onto a mutex and got stuck, and whether a lower level user could undo that.