Add concurrency test + Installation Validation
This implements validation of installs using hostfxr apis to ensure the install actually works and not just that the manifest is tracking the installs correctly in e2e tests.

It also adds a test to show that we can do multiple installs in the same directory without failing.

It also improves the existing test logic to not assume a hard-coded debug value for the dnup process.