Enable preview SDK in Test run
this is on .cmd because the claim is the release version of vs 2022 doesn't allow preview sdks. the global.json should also be able to enable preview sdks but many tests use a custom global.json file.

Whether this will actually resolve correctly, I don't know. But if the only problem was that it could resolve before but now skips preview sdks, then this could work.