Remove razor_tooling from format example ci test
As discussed in https://github.com/dotnet/sdk/pull/51376, dotnet format launches another process for doing MSBuild invocations; the issue is we're trying to talk on the named pipe to that process and it just ... never connects. This appears to be an issue thats consistent across main and 10.0.1xx at the very least and only in razor_tooling.

It might be due to some workspace configuration for razor. The dotnet-format team is working on a PR, but for now to unblock we will do this.