Set MSBuildSDKsPath
Based on the code in msbuild for the sdk resolver, this seemed to be what took effect the most to me.

I think \r is likely correct, considering we also set DOTNET_SDK_TEST_MSBUILDSDKRESOLVER_FOLDER to the \r folder.

What is surprising to me is that, that is not working. The dotnet on the PATH is also found but not by msbuild. I assume that this may mean it's VS who is rejecting the SDK, which would mean an environment variable won't be enough unless it's able to turn off the preview sdks disable check. I'll try to find the code for that or get the enable preview sdks bit to work inside of helix