Merged PR 55868: Configure SB leg to use container
The source build leg in the CI pipeline is failing in the `Generate delegation SAS Token` step:

```
##[error]Azure CLI 2.x is not installed on this machine.
##[error]Script failed with error: Error: Unable to locate executable file: 'az'. Please verify either the file path exists or the file can be found within a directory specified by the PATH environment variable. Also check the file mode to verify the file is executable.
```

This is caused by switch in the distro version of  the agent from Mariner 2.0 to Azure Linux 3.0 and the Azure Linux 3.0 image doesn't contain the Azure CLI. However, the job should really be using the SB container image, not directly on the build agent.

These changes fix the configuration to run the job in a container. It was misconfigured before because the image needs to be defined as a pipeline resource.

----
#### AI description  (iteration 1)
#### PR Classification
Configuration update to utilize containers for the SB leg in the CI pipeline.

#### PR Summary
This PR restructures the CI configuration by defining a container alias and updating its usage in the platform settings.
- `.vsts-ci.yml`: Added a `containers` section with the `sourceBuildManagedContainer` definition.
- `.vsts-ci.yml`: Modified the platform configuration to reference `sourceBuildManagedContainer` instead of a direct image URL.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->