Manual Edits - Fix Completions and XLF
Fixes bugs from copilot that makes the tasks work.
For evidence of this working and a demo, please see my fork which allowed me to test the actions by pushing to my own main, without running in prod. The most recent comments as of now which reflect the current code working are https://github.com/nagilson/sdk/pull/4#issuecomment-3647980596 and https://github.com/nagilson/sdk/pull/7#issuecomment-3647969792

https://github.com/nagilson/sdk/pull/4

Adds /completions and /xlf as shorter aliases.

Also remarks on /backport in the dev guide - it's not directly in the SDK but we use it a lot.

It might be worth porting to arcade for the xlf one :)